# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "FAQ Bot",
        "description": "L5 Swagger API description",
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/questions/{questionId}/answers": {
            "post": {
                "tags": [
                    "Answers"
                ],
                "summary": "create a  answer ",
                "description": "Create a an answer for a specific question",
                "operationId": "create answer for  question with given Id",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "Id of question",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer TOKEN",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "answer",
                        "in": "formData",
                        "description": "answer text",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "succes: true"
                    },
                    "422": {
                        "description": "The given data was invalid"
                    }
                }
            }
        },
...
    },
...
}